Struct stark::graphics::Engine

pub struct Engine<B: Backend> {
    clear: [f32; 4],
    lines: Vec<[f32; 4]>,
    colors: Vec<[f32; 3]>,
    buflen: u64,
    state: GraphicsState<B>,
    should_configure_swapchain: bool,
}

Sail-specific graphics engine state

Fields
clear: [f32; 4]lines: Vec<[f32; 4]>colors: Vec<[f32; 3]>buflen: u64state: GraphicsState<B>should_configure_swapchain: bool
Implementations

impl<B: Backend> Engine<B>

fn new(state: GraphicsState<B>) -> Self

Initialize the graphics engine

fn set_clear(&mut self, clear: [f32; 4])

Set the clear (background) color

fn add_line(&mut self, points: [f32; 4], color: [f32; 3])

Add a line, with two endpoints and a color

fn empty_lines(&mut self)

Empty the engine of all lines

fn state_pipeline_setup(&mut self)

Set up an appropriate graphics pipeline for the engine

fn state_buffer_setup(&mut self)

Acquire memory and create buffer for vertex data

fn buffer_size_check(&mut self)

Check whether the buffer has enough space for all vertices

fn setup(&mut self)

Prepare the engine to accept draw calls

fn draw_frame(&mut self)

Draw a single frame according to the engine state