Struct stark::graphics::GraphicsState

pub struct GraphicsState<B: Backend> {

    surface_extent: Extent2D,
    instance: B::Instance,
    surface: Option<B::Surface>,
    adapter: Adapter<B>,
    device: B::Device,
    queue_group: QueueGroup<B>,
    surface_color_format: Format,
    render_passes: Vec<B::RenderPass>,
    framebuffer: Option<B::Framebuffer>,
    command_pool: Option<B::CommandPool>,
    command_buffers: Vec<B::CommandBuffer>,
    vertex_memory: Vec<B::Memory>,
    vertex_buffers: Vec<B::Buffer>,
    pipeline_layouts: Vec<B::PipelineLayout>,
    pipelines: Vec<B::GraphicsPipeline>,
    submission_complete_fence: Option<B::Fence>,
    rendering_complete_semaphore: Option<B::Semaphore>,

}

Stores all persistent gfx-hal (Vulkan) objects

Fields
surface_extent: Extent2Dinstance: B::Instancesurface: Option<B::Surface>adapter: Adapter<B>device: B::Devicequeue_group: QueueGroup<B>surface_color_format: Formatrender_passes: Vec<B::RenderPass>framebuffer: Option<B::Framebuffer>command_pool: Option<B::CommandPool>command_buffers: Vec<B::CommandBuffer>vertex_memory: Vec<B::Memory>vertex_buffers: Vec<B::Buffer>pipeline_layouts: Vec<B::PipelineLayout>pipelines: Vec<B::GraphicsPipeline>submission_complete_fence: Option<B::Fence>rendering_complete_semaphore: Option<B::Semaphore>
Implementations

impl<B: Backend> GraphicsState<B>

pub fn new(window: &FrameHandle, name: &str, width: u32, height: u32) -> Self

Initialize the graphics system and track necessary state

pub fn draw_clear_frame(&mut self, color: [f32; 4]) -> Result<(), &str>

Draw a frame that is cleared to the specified color

pub fn config_swapchain(&mut self)

Reconfigure the swapchain (usually due to new surface extent)

pub fn set_extent(&mut self, width: u32, height: u32)

Set the size of the 2D graphics surface, in pixels

unsafe fn make_buffer(
&mut self,
buffer_len: u64,
usage: Usage,
properties: Properties
) -> (B::Memory, B::Buffer)

Create a new buffer for graphics processing and bind its memory

fn compile_shader(glsl: &str, shader_kind: ShaderKind) -> Vec<u32>

Compile GLSL shader code into SPIR-V

unsafe fn make_pipeline(
&mut self,
pipeline_layout: &B::PipelineLayout,
vertex_shader: &str,
fragment_shader: &str,
primitive_type: Primitive
) -> B::GraphicsPipeline

Generate a basic graphics pipeline