Module stark::sail::eval

Structs

EvalStack
Sail evaluation stack

_Frame
Contents of a frame head in this Sail evaluation stack

Enums

FrameOffset
Offsets into an evaluation stack frame (header included)

Opcode
Operation code for the frame; stored as a tag at the frame start

Functions

eval
Evaluates a Sail expression in a freshly created stack